multi.c: inotify callback support for deferred client-connect
Do not generate data channel keys in inotify callback if
deferred client-connect hasn't complete - this is the case when
callback is called for deferred auth and deferred-cc is enabled.
This is because NCP negotiation is performed after completion of
deferred client-connect.

Do not generate data channel keys if push request is not received -
in this case NCP won't be performed, since we haven't got cipher list
from the client.

Without NCP, weak BF-CBC cipher is used.

Signed-off-by: Lev Stipakov <lstipakov@gmail.com>