Fix condition to generate session keys
For --nobind clients OpenVPN reuses the context and tls_multi structs
of the previous clients and does not rerun the connect scripts on
connect. But since it is a new client connection, the key_id is 0 and
we postpone the key generation but it will never happen.

This commit changes postponing the key generation to the right
condition of NCP done for this session.

Signed-off-by: Arne Schwabe <arne@rfc2549.org>