Add detailed man page section to setup a OpenVPN setup with peer-fingerprint
This is meant to give new users a quickstart for a useable OpenVPN
setup. Our own documentation is lacking in this regard and many often
tutorials that can be found online are often questionable in some
aspects.

Linking the invidiaul RST file on github also give a tutorial
in a nicely formatted way.

Signed-off-by: Arne Schwabe <arne@rfc2549.org>