Modernise OpenVPN defaults and introduce '--compat-mode'
TLS 1.0 should be allowed anymore in a sensible default configuration. Bump
the default to TLS 1.2
Also modify --cipher not to be automatically appended  and default
allow-compression to no. Thisalso allows a default configuration to be
compatible with DCO.

Also introduce --compat-mode version to allow an easy way for UI/users
to maximise compatibility to earlier versions at the cost of DCO and
security.

Signed-off-by: Arne Schwabe <arne@rfc2549.org>