Fix management password prompt with redirected log
When management interface is password protected, openvpn
writes "Enter Management Password" to stderr. As it turns out,
when log is redirected to a file, that prompt is also written to that
file and not to the "original" stderr. Moreover, on recent Insider build
(21390.2025) openvpn exits with fatal error

  get_console_input_win32(): unexpected error: No such device or address (errno=6)

while attempting to write that prompt.

When redirecting stdout/stderr, we use _dup2() to associate stderr
descriptor with the log file. This call closes file associated
with stderr descriptor, which might explain why it has stopped
working (original stderr is closed and WriteFile() fails) and on
current versions it appears to work "by accident" - not failing
but use redirected stderr instead of original one.

I played a bit with DuplicateHandle() for STD_ERROR_HANDLE
but got the same behavior.

Fix fatal error on Insider build by always writing prompt
to the new stderr and removing concept of "orig_stderr".
For current Windows releases behavior doesn't change.

Signed-off-by: Lev Stipakov <lev@openvpn.net>